libquest: Add method to get the loaded episode from Registry/Quest
It useful to be able to know what is the currently loaded episode, so
these changes add a method to the Registry and to the Quest in order
to retrieve that.
The reason why we add a method to the Quest as well, instead of just
relying on the Registry's one, is that in the Quest case it is the
episode loaded at the time the quest is instantiated (i.e. the Quest's
episode) and not the most recently loaded episode, as is the case with
the Registry. Usually both values should be the same, but this way we
avoid a case where a quest may use that value while the Registry has
just loaded a new episode, so this approach is mostly a defensive
measure.

https://phabricator.endlessm.com/T26207