libquest: Make sure Quests set the correct episode as complete
When a Quest uses its complete_current_episode, it is in fact trying
to set its own episode as complete, and not the current episode in the
GSS. While those should be the same, we've recently run into a problem
where the next episode would be set just before a quest sets the
current episode as complete. So, even if the order of those calls is
a mistake, this patch adds a check to the complete_current_episode to
ensure the Quest's episode is the one currently set in the GSS, before
completing it, thus avoiding such future problems.

https://phabricator.endlessm.com/T26207