quests: Complete episode before setting the next one in BreakingIn
The next episode is being set up in BreakingIn's complete property
setter, but the call for completing the current episode is being called
after the quest is set as complete; this results in setting the next
episode before completing the current one. Due to previous changes in
the way an episode is set as complete, the problem above no longer
results in the next episode being set as complete, but we should still
make sure that the order is the correct one.

https://phabricator.endlessm.com/T26207