clubhouse: Stop any delayed messages if the quest is finished
Quest messages that are not a result of user interaction can be delayed
for a few milliseconds with the goal that the user is better aware of a
change in the message popup dialog.
However, if a quest sends one of those messages right before it is
stopped, then the message would still be shown after that delay (when
the quest is not even running anymore).

This patch fixes this issue by stopping an existing delayed message when
the quest is finished.
As an extra effect from this patch, if two messages are sent one right
after the other, instead of scheduling and showing the two messages
(which means the first message would be visible for just a few
milliseconds), the second message just stops the first one from ever
being shown which makes for a more solid UX.

https://phabricator.endlessm.com/T26229