clubhouse: Show episode poster also when an episode is started
If a quest sets the current episode as complete and immediately advances
to the next episode, then the episodes page (that listens to GSS changes
and shows the episode teaser when needed) may not be aware of the
current episode completion, since when it reads the values from the GSS,
the new episode will be already set.

Thus, in order to avoid episode posters from not being shown when the
conditions above happen, these changes also show the episode poster when
a new episode has been set up in the GSS and the previous episode's
poster hadn't been shown.

https://phabricator.endlessm.com/T26228