libquest: Simplify how Quests can complete/advance episodes
Quests who set episodes as complete or advance to a next episode need
to reimplement their "complete" property setter, but that means more
things to verify (e.g. making sure the base implementation is called)
and therefore is more prone to errors.

These changes add those changes to the base Quest class's "complete"
property setter, and allow to control the completion or advancing
of an episode by simply overriding the __complete_episode__ and
__advance_episode__ variables.

The use of these variables by the quests will code at a later commit.

https://phabricator.endlessm.com/T26228