libquest: Add the notion of next episode in the Register
Currently quests are setting the next episode (to finish the current
one) by directly saving it using the GSS. While this guarantees that
the quest advances the episode to the one it's exactly requesting, this
was a temporary measure as the notion of "next episode" is a more global
one, and thus we should having it in a more global context.

For that reason, these changes set the NEXT_EPISODE to the episodes'
module, and load it into the Registry as well, oferring it by a new
method "get_next_episode_name". So this will be used a t a later commit
to make the transition to the next episode more transparent to quests.

https://phabricator.endlessm.com/T26228