libquest: Use the next-episode name automatically in the Quest class
Currently quests are setting the next episode (to finish the current
one) by directly saving it using the GSS. While this guarantees that
the quest advances the episode to the one it's exactly requesting, this
was a temporary measure as the notion of "next episode" is a more global
one.

Since now we can get the next episode name straight from the Registry,
these changes allow a Quest to use that name when setting the next
episode in the GSS, and also allow for the new __advance_episode__ to
be a boolean variable instead.

This means that if quests want the episode to be advanced when they're
completed, they should just declare "__advance_episode__ = True" and
the rest is done automatically.

https://phabricator.endlessm.com/T26228