quests: Refactor the quests' episode completion/advancing
Now that there's a much simpler way of completing episodes or setting
the next one, we should make sure that the quests use that. So this
patch changes that logic in the quests and also adds the information
about what the next episode is in each episode that needs it.

https://phabricator.endlessm.com/T26228