Move the highlight animation logic into the QuestSet
The logic for setting up the highlighted animation (called "hi") stopped
working after started using the body-animation property for setting the
Character's animation with the same name.

These changes move that logic into the QuestSet, so both the highlighted
and body-animation properties are managed in the same place and this
also fixes the regression mentioned above.

https://phabricator.endlessm.com/T26167