clubhouse, libquest: Use all the message information to offer a quest
If there is a QUESTION message in the catalog, it should be used to
build the message that offers the quest. However, only the text and
SFX sound were used until now. The character was set to the main one
from the quest or quest-set. And the character's mood wasn't falling
back to the default mood, it was using the last mood used.

This change makes the initial message use all the information
available. Except for the background sound, which by design is
ignored.

Also, if a character is set in QUESTION, it will overwrite the Quest main character.

https://phabricator.endlessm.com/T26237