apps: Prevent exception in Lightspeed.set_level
If we try to set the level in the Lightspeed app but the app is not
running, then the method returns None, meaning we cannot compare that
result with a number (resulting in an uncaught exception).
To avoid that we just check that return value and early return if
needed.

https://phabricator.endlessm.com/T26283