Add initial support for unit tests
These changes add the support for unit tests to the Clubhouse and a
first unit test (just checking if the default episode is loaded).

Particularly these changes consider all Python files starting with a
"test_" prefix in the "tests" folder as tests to be built.
The built tests are run after installing the Clubhouse in the Flatpak
as they need the files in they're final condition (right path + the
generated config) for the tests.

By default the tests are turned off in the base manifest file, but
they're enabled when using the build-local-flatpak.sh script so they're
seamlessly run during the normal development.

As part of adding support for tests, we need to make sure that the tests
run as close as possible to the real Clubhouse internal use, but without
affecting the user's game state.

Thus, these changes also add a new mock class to be used by the
GameStateService for testing. This class keeps all the changes in
memory.
Maybe using a DBus mock service would be a more elegant solution, but
that would take longer than this implementation, and thus let's keep
that as a possibility to be considered later.

https://phabricator.endlessm.com/T26337