libquest: Automate how the QuestSets' use the empty message
All of the QuestSets were overriding the get_empty_message method and
fetching the actual message from the strings catalog.

Thus, these changes make that generic use easier by using the string
catalog in the base QuestSet class implementation. And still defaulting
to the __empty_message__ class variable if the mentioned strings are not
available (for now).

https://phabricator.endlessm.com/T26281