libquest: Use episode specific empty messages when needed
So far, the messages for when a character has no quests to offer to the
user were defined globally, but a recent request from the quest writers
is to have this message per episode.

Hence, these changes look for messages with a NOQUEST_ + episode-ID
before using the global ones (simple NOQUEST prefix) for the empty
messages.

https://phabricator.endlessm.com/T26281