Implement initial redesign of Episodes page
The Episodes page has a new design and UX, and these changes add the
initial support for that. Namely, they replace the image-based Episode
view by a listbox that displays all the episodes + their badges.

Not all the Design specs have been implemented yet, as this is just the
base for doing more specific work that requires having this view instead
of the old one.

https://phabricator.endlessm.com/T26393