flatpak: Force gtk notification backend
The new version of xdg-desktop-portals filters notifications with
GFileIcon and we're using that to send the character animation path to
the shell.

This change will force the use of the gtk notification backend in glib
so we'll avoid the portals restriction for notifications and everything
will continue working as expected.

https://phabricator.endlessm.com/T25823