clubhouse: Don't show the description on not-started episodes
Per Design, in the Episodes page only episodes that are current or
completed should show their description when clicked.
For that reason this patch only sets up the description label and
connects the button's click to reveal it on the mentioned episodes.
The buttons for episodes that have not been started yet are also set
as insensitive so the button doesn't change its style if interacted
with.

https://phabricator.endlessm.com/T26393