libquest: Add Registry method to retrieve the current progress
The new Episode page design needs to show badges for the current
episode that show the progress of the user so far, i.e. the percentage
of completion for the current episode.

For that, these changes add a new method to the Registry that allows to
retrieve that progress in a 0 to 1 fraction.
It also adds a convenience method to retrieve all the quests for the
episode.

https://phabricator.endlessm.com/T26393