clubhouse: Update the current episode's percentage when needed
Now that we have a way to easily retrieve the current episode's
progress, we need to update the percentage for the Episode object used
by the Episodes page.
This patch does that by setting up the percentage according to the
progress initially after creating the current episode row, and also
when a quest is finished running (and is completed).

These changes create a dictionary of the episode rows which can be
redundant since we don't need constant access to all episodes but we
can simplify this at a later stage, instead of growing this patch more.

https://phabricator.endlessm.com/T26393