clubhouse: Reload the Episodes page when a new episode is loaded
When the Clubhouse transitions to a new episode, the Episodes page needs
to reflect that (i.e. set up the completed + new current episode row
style). For that, this patch just reloads the whole page when a new
episode is loaded.

The ideal way of doing this would be to just restyle the previous +
newly loaded episode rows, but that involves connecting to an actual
episode load, and that's not possible yet, so it's left for later.

https://phabricator.endlessm.com/T26393