clubhouse: Continue playing the clicking an episode's progress badge
When the player clicks the badge of the current and incomplete episode
(play or progress badge), then the Clubhouse should focus the main view
and use the first active character to propose a quest to the user when
there no running quest. If there is a running quest, then the quest's
character is the one used (will ask the player to whether continue/stop
the quest).

These changes do that and also move the logic for showing the poster
from the EpisodeRow into the EpisodesPage so we don't end up with the
click behavior split across two objects.

https://phabricator.endlessm.com/T26393

fixup! clubhouse: Continue playing the clicking an episode's progress badge