clubhouse: Fix initial placement of the Episodes badges
The Episodes page places the episodes' badges on top of the episode
rows it may happen that the badges are requested to be placed before the
badges box widget is realized, which results in a failure to translate
the coordinates between widgets.

To solve the issues mentioned above, these changes ensure that the
badges are only placed when the badges box widget is already realized.

https://phabricator.endlessm.com/T26393