clubhouse: Use the default application in the ClubhousePage
The ClubhousePage uses the application from its app window and asserts
that it is valid. Instead of doing that, it is more efficient and safe
to just use the default GApplication for the process (so if e.g. in the
future the window has no application attached to it, this code will
still work).