libquest: Allow Quests to wait for toolbox topic clicked
Adds an async connection to the clicked signal of the passed
ToolboxTopic. Just like the confirmed step, it stores the clicked
toolbox topic (app name and topic name) in a private
property. Upcoming quests have to wait for topics being clicked, and
act accordingly depending on which topic was clicked.

https://phabricator.endlessm.com/T26514