Fix loading and finishing demo episode
The demo episode needs to be in the episodes info file and have its
badges in order for it to work correctly after the Episodes page
redesign.

These changes thus add the needed info to the episodes CSV file and
just copy + rename the episode1's badges. No episode poster for demo
exists, so that still needs to be fixed but it'll be done some other
time.

https://phabricator.endlessm.com/T26393