quests: Ensure BreakingIn is not left in a blocked state
Since the BreakingIn trap first sets the state so Riley is trapped, and
only then it proceeds to set the quest as complete. It may thus happen
that the quest is left incomplete but with a trapped Riley (running out
of battery, crash, etc.), which means the user can no longer start that
quest and thus cannot finish the episode after that.

This patch prevents that by detecting such a case (incomplete quest +
trapped Riley) when the quest is initialized or when it's finished, and
resets Riley's state so the user can start the quest again.

https://phabricator.endlessm.com/T26578