libquest: Only give an item if the quest is not cancelled
Sometimes quests give items after using a "wait function" and without
checking whether the wait function succeeded or not. So if the quest is
cancelled, the wait function finishes the wait and thus any item given
after that is successfully given.

In order to prevent that, this patch ensures that the give_item method
first checks for whether the quest is cancelled, in which case it
doesn't give the item.

https://phabricator.endlessm.com/T26451