Use real percentages for the current episode's progress badge
Even though we calculate the real percentage of completion for episodes,
that number was getting filtered to match one of the percentage badges.
Since there were only a few percentage badges available, this meant that
there wasn't usually a clear step in the percentage shown to users after
completing a quest.

This patch changes that by using progress badges for just the progress
hexagon (they don't have any percentage in them), and display a label
with the real percentage of episode completion on top of it.

https://phabricator.endlessm.com/T26630