quests: Fix delayed messages in MazePt1
The quest blocks on the wait_confirm() messages as expected. So the
step step_play_level should wait for either app property changes or
the message confirmed. To keep the desired implementation, keep track
of the messages already confirmed, and display the last one not yet
confirmed. Also, use the new dismiss_message() in the levels that
don't have a message to confirm or hint.

https://phabricator.endlessm.com/T26641