quests: Fix detecting the enemy direction in LightSpeedEnemyA2
Split the playing between A) waiting for the spinner to spawn and B)
checking if the spinner is going up. To check if it's going up, the
initial min/max values reached by the spinner are stored.

Also check if the user flips the app while playing, and act
accordingly.

https://phabricator.endlessm.com/T26639