quests: Fix race condition disabling tools in FizzicsKey
The app now exposes a loading level property, so the quest can wait
for it and only then disable the tools.

Also, add a small delay before disabling non-rocks.

Also, calling reset is not needed. It just resets the level properties
to the initial values, and setting the level already does that.

https://phabricator.endlessm.com/T26699