T26694 levifix (#571)
* quest: Improve detect progress in LeviHackdex

Wrap step_detect_progress with the decorator to monitor the app is
launched. There is no need to check if the app is running inside the
step, because it comes directly after launching it in the step_begin.

https://phabricator.endlessm.com/T26694

* quests: use Quest.is_panel_unlocked in LeviHackdex

* quests: Check for success earlier in LeviHackdex

https://phabricator.endlessm.com/T26694

* quests: Rename variable in LeviHackdex

From FLIP_COUNTER to RESTART_COUNTER, because users can flip the app
and flip it back without doing any changes, and in that case the app
won't restart.

https://phabricator.endlessm.com/T26694

* quests: LeviHackdex, don't monitor app launched after success

The step_part2 happens after the player reached the goal. If the
player flips the app at this stage and changes the code, the app will
quit and restart and the quest will abort.

https://phabricator.endlessm.com/T26694

* quests: LeviHackdex, increase the timeout waiting for the app restart

It doesn't hurt to wait for 10 seconds for safety.

https://phabricator.endlessm.com/T26694