tools: Restart the GSS in the build-local-flatpak script
The GSS now has a longer timeout so it can still be alive after a new
Clubhouse build is finished. This may impede a fresh test of the
Clubhouse with an updated GSS because the service may still be running
with the older version.

Thus, we simply ask the GSS to reload its conf every time we build the
Clubhouse locally.

https://phabricator.endlessm.com/T26730