quests: monitor level change in every wait of FizzicsKey
Everytime the quest waits for a message to be confirmed or a property
of the app to change, it also checks if the level changed. If the
level changed, it decided what to do.

- If the intro cutscene wasn't played, it plays it.

- If the level is outside of the range this quest handles, the quest
  starts again.

- Otherwise it goes to the corresponding level.

This could be better implemented with a python decorator similar to
the @Quest.with_app_launched, but I ran out of time.

https://phabricator.endlessm.com/T26688