quests: Display right messages when the level changes in RileysLevels
When the level is changed in the RileysLevels quest, unless they
were really the 12th-16th levels, the message for the levels was not
changed.

This patch simplifies how the quest's code is structured and ensures
that, if the level is changed, the message for that level is displayed
again.

https://phabricator.endlessm.com/T26280