quests: Fix detecting a level change in RileysLevels
When a ball dies in Fizzics, the RileysLevels quest checks for when this
state changes. If that state changes because a new level is about to be
loaded, then the quest will fail to detect the new level change because
of the timing between the ball state and the level changes, i.e. the
quest checks for the level right after the ball state changes, which is
too early.

Thus, in order to give time for the level to change and for the quest to
correctly detect it, we pause for 1/4 of a second after the ball state
is changed. This pause is hardly perceivable by the user, and is enough
to detect the level changes correctly.

https://phabricator.endlessm.com/T26280