tools: Change how the git pre-commit hook is installed
The git pre-commit hook was simply being copied to the right path when
setting it up. This had two problems:
  1. If a pre-commit hook existed, then it'd be overwritten, which could
     not be what was desired by the user;
  2. If a the pre-commit hook was updated, then it'd not be used until
     setting it up again.

The two issues above can be simply solved by setting up the pre-commit
hook as a symlink to the hook file we ship, thus it will fail to set up
if a pre-commit hook already exists, and it will be up to date just by
pulling the latest changes from the repo.

https://phabricator.endlessm.com/T26750