build: Run lint as part of the build process
So far the support for lint existed in the Clubhouse but was being used
only directly or by a git pre-commit hook, so it depended on the user to
run it, and as a result we had several PRs that were merged with lint
issues (not only "innocent" coding style but also uses of modules
without importing them, etc.).

Now that we have CI for the Clubhouse, it is important to add the lint
checks to the build process, so they are also run for every PR and catch
errors earlier. At the same time, if whoever makes a patch also builds
the Clubhouse, they will be running the lint automatically so there's no
need to remember to run it even locally.

https://phabricator.endlessm.com/T26750