tools: Skip the lint on local builds if the pre-commit is set up
Developers who regular work in the Clubhouse ought to have the git
pre-commit set up to perform the lint on every commit, so having the
lint check on every local build may be redundant and slows up a build
process that's frequently run.
In order to avoid this, the build-local-flatpak.sh script now checks
whether the mentioned pre-commit hook is set up, and if so it just
turns off the lint option from the build, thus making the build
quicker.

This means that the chances of filing a PR with lint failures after
building/committing locally are very low.

https://phabricator.endlessm.com/T26750