libquest: Make Registry.get_current_quests return a dictionary
The mentioned method returns a list, but it's very useful sometimes to
be able to get a Quest object directly by its id, so this patch changes
the return of that method to be an OrderedDict (so we keep the same
order as we had in the list), and updates any caller's code
appropriately.
The usefulness of these changes will be more evident in coming commits.

https://phabricator.endlessm.com/T26812