Add --set-quest CLI option
Very often, for development or testing, it is very hard to be able to
try a random quest, as it requires one to solve all quests up to that
one, or to fake it by tweaking the GSS which takes time.

Thus, these changes introduce a new CLI option --set-quest that sets up
the given quest as available, automatically completing all the quests
in its dependency graph. The episode is reset as well, i.e. even if
we're setting a quest from the current episode, the GSS will be reset
as this avoids e.g. having quests after than the desired one still as
complete, or keys that shouldn't be set before the desired quest.

https://phabricator.endlessm.com/T26812