libquest: Allow Quests to declare items to be given on completion
Now that we have a way to quickly set up the state of an episode as if
played until a certain quest (with the --set-quest CLI option), even
though the quests get set as complete automatically, there is no way to
set up any important items that the quests may give (such as keys).

These changes introduce the new Quest class variable
__items_on_completion__ which is used to set up any items that need to
be in the global state when the quest is completed. Thus, any items or
keys that are not given optionally to the user during a quest, should be
declared in this variable.

https://phabricator.endlessm.com/T26812