libquest: Add an optional keyword argument to Quest.give_item
The items that a quest gives to the user should be pre-declared in the
quest. However, there may be keys that are optional, which in that
case they shouldn't always have to be pre-declared. Thus, this patch
adds a new keyword argument "optional" to the Quest.give_item in order
to be able to give items that are not pre-declared.

https://phabricator.endlessm.com/T26812