tests: Add define_quest convenience method
Some tests need to define new Quest objects from scratch, so this patch
adds a new method to the clubhouseunittest module for defining a quest
just by choosing its id, character id, and dependencies.

In these changes the use of the PhonyQuest in test_quests in also
replaced by this method.

https://phabricator.endlessm.com/T26812