libquest: Delay QuestSet registration
Instead of instantiating the QuestSet right when its module is loaded,
this patch delays the actual instantiation + addition to the list of
quest sets so the QuestSet objects are instantiated only after the whole
episode folder is loaded. This is important for upcoming changes in
which we want to be able to include Quests in a QuestSet without having
to import them directly before.

https://phabricator.endlessm.com/T26875