tests: Don't use mock in the test_questsets
In the test_questsets we were mocking the mentioned method in a
permanent way, i.e. the list of quest sets in the Registry would always
return the mocked values, even when running tests after the one where it
was mocked.

This could be fixed by using mock.patch for the needed functions, but
since we are talking about objects in memory and the tests don't run in
the same process as a regular Clubhouse use, it's just simpler to change
the Registry object directly so we avoid problems like this in the
future.

https://phabricator.endlessm.com/T26875