libquest: Allow to set QuestSets' quests without importing them
QuestSets need to import their quests before being able to set them up
in their internal list. This means that for some QuestSet files, the
number of lines just importing quests is greater than the actual
QuestSet's code.

To make this easier, this patch allows to set up the QuestSet quests as
strings (using their ID). Currently both approaches are supported and
even when mixed (strings and classes set up in the QuestSet's
__quests__).

https://phabricator.endlessm.com/T26875