libquest: Deprecate Quests' name constructor argument
This name is only being used for printing and is either very similar
or equal to the quests' class name, so these changes make it not be
used anymore (the get_id() is now used for the __repr__) and print a
warning if the argument is used.

The main motivation for this is to simplify how quests are defined
in the near future (they should not have to pass so many arguments
not define a constructor even).

https://phabricator.endlessm.com/T26875