libquest: Set the proposal_message_id as a class variable in Quest
We have been using a set of special class variables in the Quest class
for overriding some of the default behavior for the quests, so instead
of passing the proposal_message_id as a constructor parameter, quests
should also use a special variable for that.

This patch adds that variable as __proposal_message_id__ as uses its
value if the constructor's related parameter is not set. So for now both
solutions coexist but the idea is to drop the constructor parameter once
quests stop overriding their constructors.

https://phabricator.endlessm.com/T26875