libquest: Make the Quest receive its QuestSet on construction
Quests didn't use to need a QuestSet initially, so they didn't receive
it as a construction parameter but they were simply set up right after a
Quest object was initialized. However, for a while now the quests do
need to be set in a QuestSet in order to be offered to the user, and the
Quest class has methods that count on having a QuestSet, so the QuestSet
should really be passed as a construction parameter for the Quest.

https://phabricator.endlessm.com/T26875