tools: Check items files in check_strings script
We have recently missed updates to the "episodes" spreadsheet, so we
should make sure they are checked by the check_strings script just like
the quests' strings.

This patch adds those extra checks and refactors the script a bit too,
so it should be easier to add more checks if needed in the future.

https://phabricator.endlessm.com/T26902