tools: Don't abort the commit when check_strings fails
The pre-commit hook was aborting the commit if the check_strings script
failed, but we've come to realize that, since this was annoying for the
developers, some would just turn off the check.

Thus, in order for everyone to realize when string/csv files do need to
be synchronized but without the check getting in the way, this patch
just lets the check print the warning without exiting with a bad status,
so the commit proceeds as usual. The lint check remains as previously,
i.e. a lint fail prevents the commit (and this is needed).

https://phabricator.endlessm.com/T26902