clubhouse: Hardcode the window's height
Now that the Clubhouse window is of a normal type, we need to make sure
that it is not resizable by the user. That also means we cannot keep
setting its height based on the screen and reacting to monitor changes,
etc. So these changes just hardcode the window's height and make it
unresizable for now.

https://phabricator.endlessm.com/T27186