quests: Personalize desktop after the FirstContact quest
This quest enables the hackmode, but only when it's resolved, so if it's
not completed we should go back and disable the hack-mode. The hack-mode
desktop personalizations are applied after resolve the puzzle.

The hack-mode should be enabled at the begining because in other case
the FtH button won't be there, but if the puzzle is not finished we
should disable again.

https://phabricator.endlessm.com/T27152