quests: Avoid launching Unlock app twice
The DBUS call to get the app property triggers the app launch if the app
is not running. We don't want this because the purpose of querying these
properties is only for reading, if the app is closed during a step we
should avoid running the app again so every function that gets a
property shoud check the app running status.

https://phabricator.endlessm.com/T27231