clubhouse: run autorun on activate
Instead of launch the autorun quests only in handle_local_options, this
patch moves that to the activate, to launch the FirstContact quest if
it's not finished everytime the clubhouse is activated, just before
showing the window.

The clubhouse won't be launched anymore outside the user session, so we
can remove all the session check done in the autorun.

The window will be shown only if there are no autorun quests to run.

https://phabricator.endlessm.com/T27232