Allow the Clubhouse window to slide down from the top
This commit reverts part of the work done in commits 22ce4b9 and
d13293c.

Now the window is positionated to the top of the screen and has got
undecorated (window borders are removed) so the Shell window
manager can apply a slide down effect to the Clubhouse window
imitating the Discovery Feed behavior.

https://phabricator.endlessm.com/T27162