system: Use gtk_app proxy to launch applications
With the flatpak that combines the clubhouse and the HackUnlock
application we don't have the com.endlessm.HackUnlock.desktop file
exported so the usual call to launch this application won't work.

This patch tries to use the org.gtk.Application dbus call to launch the
application if the usual method fails. We're exporting correctly the
HackUnlock dbus as an activatable service so this interface is there and
should do the work.

https://phabricator.endlessm.com/T27274