data: Add X-Endless-HackShader key to desktop file
When using the katamari flatpak that combines the HackUnlock and the
Clubhouse apps, the shell is using the com.endlessm.Clubhouse.desktop
file for the HackUnlock app, so to apply the filter to the HackUnlock
toolbox this is needed.

This shouldn't affect the clubhouse itself because we don't have toolbox
for the clubhouse.

https://phabricator.endlessm.com/T27274