system: flatpak overrides for App
To be able to automatically export a useful dbus interface for 3rd party
apps we need to use the libclippy in the GTK3_MODULES. This new method
allow us to override the flatpak environment for any application so we
can inject the clippy lib in flatpaks without modify the manifest
itself.

The enable_clippy method can be called to add the override file or to
remove it.

This method can be called in the future from a quest or maybe when the
hack-mode is activated, with all apps that we want to make hackables.

https://phabricator.endlessm.com/T27274