Allow quest dialogues inside the clubhouse
These are going to support the narrative quests when the character
mission view is implemented. For now these dialogues appear in the
clubhouse main view.

The default is to keep all dialogues as pop-up notifications in the
shell. Quest authors can pass narrative=True to any quest API
displaying dialogues.

https://phabricator.endlessm.com/T27307