clubhouse: Rename show_message method and move it to Message class
We will need to show messages in the character view, but currently
the show_message method is dependent on the ClubhouseView class.
This commit moves the show_message method to the Message class
renamed as update receiving the same options as of the dictionary
passed in the 'message' signal of the Quest class.

https://phabricator.endlessm.com/T27406