clubhouse: Fix background setting on Clubhouse initialization
The previous commit caused a bug, which was to set
hack-mode-disabled-picture-uri and hack-mode-enabled-picture-uri
pointing to the same background. It gave the perspective that no
change in the background was made when switching the hack switch.

Also, due to the fact that now a change in the hack mode now
implies a change in the background and in the cursor, too, the
previous commit caused a different behavior. The previous behavior
was to show the Hack background when the HackUnlock puzzle gets
completed, but the previous commit made to show the HackUnlock
background since it started: that explanation is that the
FirstContact quest was setting hack mode to True (thus making to
change the background and cursor). This has been fixed by changing
the background and cursor mode to a non-hack mode style even when
in the hack mode.

https://phabricator.endlessm.com/T27464