clubhouse: Prevent setting style to white when selecting pathways
Each quest in the pathways view belongs to a listbox. Before this
commit the listbox was selectable and it produced to change the
style to white when you selected an item of the listbox overriding
current styles giving the sensation that text was deleted. This
commit makes the listbox unselectable to prevent that.

https://phabricator.endlessm.com/T27405