clubhouse: stop_quest only if hack_mode_enabled is false
As part of the "cleanup off state classes" the running quest is been
stopped when the hack-mode-enabled change so for the first contact quest
the quest is cancelled after launch the HackUnclock app, because the
quest is ran and the hack-mode-enabled is modified inside that quest.

This patch fix this problem doing the quest stop only if the hack mode
is disabled.

https://phabricator.endlessm.com/T27478