katamari: Use katamari to build the flatpak
The katamari tool was in another repository but to make it easier for
the deployment and maintenance this patch copies the katamari tool here,
so after merge this we can delete the katamari repository.

The clippy extension is also here, so this tool will be able to build
that too, but the CI should build the clippy extension using the
manifest inside the clippy repository.

https://phabricator.endlessm.com/T27400