animation: Avoid errors if setting up a non-existing animation
Currently, if a quest developer adds a new mission for characters
whose animation asset 'close-up' is not available, then the
Clubhouse ends up showing a lot of traces.

This commit allows to set a non-existing animation, so in this
case the current animation is cleared as well as deleted from the
AnimationSystem registry. A practical case for clearing the
animation will be seen in the CharacterView; it avoids the
CharacterView to keep the last character animation when switching
to the character view of an animation without the 'close-up'
animation.

https://phabricator.endlessm.com/T27519