animation: Pause, sometimes, character view animation on new message
By default, closeup-talk animations will be always paused. Only
when a new dialogue is added with a character equal to the main
character, play the animation.

For this, methods pause and play have been added to the Animator
class.

Animation class has been moved upwards so we can define a default
argument default_state in the 'animate' method signature:

    ```
    def animate(self, name,
                default_state=Animation.State.PLAYING):
    ```

https://phabricator.endlessm.com/T27551