clubhouse, libquest: Quests should be able to auto-offer themselves
When they become available.

To do that the schedule-quest signal should be moved to the Registry,
instead of hooking this signal from the current running quest.

This change makes Registry a GObject, with a singleton instance that
can emit the schedule-quest signal, and connect to that signal from
other objects.

https://phabricator.endlessm.com/T27850