libquest: New method to Highlight nav buttons
This new method in the Quest base class will allow to highlight one of
the nav buttons that will go to normal state when the user navigates to
the corresponding page.

If the button highlighted is the PATHWAYS, this will be un-highlighted
when the user go to one of the pathways, the CHARACTER page.

https://phabricator.endlessm.com/T27925